chore(ci): use function executor for abs and signed div tests
